// src/app/store/content/types/departments/DepartmentsActionTypes.enum.ts
export enum EDepartmentsActionTypes {
  FINDALL = '[Departments]-find-all',
  FIND_BY_ID = '[Departments]-find-by-id',
  CREATE = '[Departments]-create',
  DELETE = '[Departments]-delete',
  UPDATE = '[Departments]-update',
  SUCCESS_GET_ONE = '[Departments]-get-one-success',
  SUCCESS_GET_MANY = '[Departments]-get-many-success',
  SUCCESS_DELETE = '[Departments]-delete-success',
  ERROR = '[Departments]-failure',
}